Introduces hard-coded temporary "race" (term TBD) controlled vocabulary
This is in response to criticism of the source descriptors shown in the dropdown filter. This replaces those values with a temporary controlled vocab. The mapping of DB values to that vocab is defined in browse_tabulator_json-processor_race-value-map.js. The working doc for that map is found in https://docs.google.com/spreadsheets/d/1fgGDVx3fwk3mybtqUciRI1iuAQON1GXDv6cX1gTTgDE/edit#gid=2009764625